requirements.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-google-genai
sentence-transformers
google-generativeai
python-dotenv